requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
blinker==1.6.2
boto3==1.28.63
botocore==1.31.63
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cmake==3.27.6
contourpy==1.1.1
cycler==0.12.1
Cython==3.0.2
dataclasses-json==0.6.1
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
extra-streamlit-components==0.1.60
faiss-cpu==1.7.4
fonttools==4.43.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.37
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipython==8.12.3
jedi==0.19.1

Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.308
langsmith==0.0.42
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.1.0
numpy==1.24.4
openai==0.28.1
packaging==23.2
pandas==2.0.3
parso==0.8.3
pdf2image==1.16.3
pdfminer==20191125
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pkgutil-resolve-name==1.3.10
plotly==5.17.0
poppler-utils==0.1.0
prompt-toolkit==3.0.39
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pycryptodome==3.19.0
pydantic==2.4.2
pydantic-core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.1
pypdf2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
s3transfer==0.7.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==2.0.21
stack-data==0.6.3
streamlit==1.27.2
streamlit-aggrid==0.3.4.post3
streamlit-authenticator==0.2.3
tenacity==8.2.3
tesseract==0.1.3
tiktoken==0.5.1
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typing-extensions==4.8.0
typing-inspect==0.9.0
tzdata==2023.3
tzlocal==5.1
urllib3==1.26.17
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.8
yarl==1.9.2
zipp==3.17.0